requirements.txt
beautifulsoup4
python-slugify
pdfminer.six
